requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
accelerate>=0.30.0
transformers>=4.39.0
streamlit==1.35.0
huggingface_hub==0.23.0
python-dotenv==1.0.1
